--- a/stresstest.c
+++ b/stresstest.c
@@ -220,6 +220,7 @@ int run_tests(int64_t *sizes, int sizes_cnt, int type) {
     TEST_SORT_H(selection_sort);
     TEST_SORT_H(bubble_sort);
     TEST_SORT_H(binary_insertion_sort);
+    TEST_SORT_H(bitonic_sort);
   }
 
   TEST_SORT_H(quick_sort);