I have placed routines for sorting networks of length <= 16 in sort.h as well as the BITONIC_SORT routine. However, the BITONIC_SORT routine is no longer a "true" bitonic sort. It calls the optimal sorting networks for length <= 16 and BINARY_INSERTION_SORT for anything larger than that.