@DrMarkS 2019-08-11
I shall have a think about your other comment but I am worried it may make the code shorter at the expense of being even more unreadable.
I am curious to know if there is a meaningful performance gain on a decent machine.