@swenson 2019-08-11
I'd say that it is basically unreadable now. It's not like a human can meaningfully change the code as-is.

I'll see if I can run some benchmarks to see how much of a performance gain there is. I'd say the biggest potential problem is in partially sorted cases -- bitonic sorting always takes the same amount of time, but some of the other small algorithms can run in linear time if the data is sorted or nearly sorted.