@DrMarkS 2019-08-16
I am failing the clang-5.0 check and I don't think that is my fault.