@swenson 2019-08-20
No worries about the clang-5.0 failure. It looks like the build package for it isn't working.

So, I did some performance testing on Core i7-8700K, a fairly fast desktop CPU, and got the following results (before is the `master` branch, percentage change is with merging in this bitonic branch). These are run for arrays of 100, 1000, and 100,000 random 64-bit integers.

```
benchmark                                       old ns/op     new ns/op     delta
BenchmarkQsort100_x86_64                        1720          1601          -6.92%
BenchmarkBinary_insertion_sort100_x86_64        1428          1505          +5.43%
BenchmarkQuick_sort100_x86_64                   664           526           -20.81%
BenchmarkMerge_sort100_x86_64                   724           464           -35.90%
BenchmarkHeap_sort100_x86_64                    1491          1467          -1.58%
BenchmarkShell_sort100_x86_64                   600           668           +11.33%
BenchmarkTim_sort100_x86_64                     983           1031          +4.90%
BenchmarkMerge_sort_in_place100_x86_64          858           902           +5.20%
BenchmarkRec_stable_sort100_x86_64              2822          2909          +3.08%

BenchmarkQsort1000_x86_64                       44987         43284         -3.78%
BenchmarkBinary_insertion_sort1000_x86_64       123687        128717        +4.07%
BenchmarkQuick_sort1000_x86_64                  23635         17279         -26.89%
BenchmarkMerge_sort1000_x86_64                  28682         20912         -27.09%
BenchmarkHeap_sort1000_x86_64                   24689         24590         -0.40%
BenchmarkShell_sort1000_x86_64                  39770         38888         -2.22%
BenchmarkTim_sort1000_x86_64                    33285         33663         +1.13%
BenchmarkMerge_sort_in_place1000_x86_64         22817         22440         -1.65%
BenchmarkRec_stable_sort1000_x86_64             108041        112953        +4.55%

BenchmarkQsort100000_x86_64                     7720985       7484739       -3.06%
BenchmarkBinary_insertion_sort100000_x86_64     961360500     992938500     +3.28%
BenchmarkQuick_sort100000_x86_64                5051258       4246585       -15.93%
BenchmarkMerge_sort100000_x86_64                5953286       4958941       -16.70%
BenchmarkHeap_sort100000_x86_64                 6009982       5993264       -0.28%
BenchmarkShell_sort100000_x86_64                8375267       8376750       +0.02%
BenchmarkTim_sort100000_x86_64                  6407344       6416885       +0.15%
BenchmarkMerge_sort_in_place100000_x86_64       5674175       5697074       +0.40%
BenchmarkRec_stable_sort100000_x86_64           22536978      23274279      +3.27%
```

`qsort` is unaffected by these changes, so we can consider it the placebo, showing that there is some variability inherent in the numbers (+-5%).

However, I'm quite impressed about the increase of performance across the board of quick sort and merge sort.

I think this definitely justifies merging this code. Thank you for your contribution.